requirements.txt
pandas==2.2.3
numpy==2.2.4
ta==0.11.0
requests==2.32.3
python-dotenv==1.1.0
APScheduler==3.10.4
python-telegram-bot==21.11.1
feedparser==6.0.11
pytest==8.3.5
pytest-asyncio==0.25.3
